Publishing volume metrics to the controller

Correcting logging object of volume status


Renaming/restructuring function and file names in volumemgr


Minor changes


Addressed review comments


Addressing review comments


Addressed review comments


Addressing review comments


Addressed review comments


Minor reordering


Added missing comment


Adding ref count in volume config/status


Fixing bug of deleting used volumes


Revert "Fixing bug of deleting used volumes"

This reverts commit 0b2e4a4f9e0276c5c3006c02234635dfe24f52a3.

Fixing decrementing refcount 0 issue
